Fix ContractGeneratorTests so that given options are used; naming strategy refactor (combine classes, leave interfaces separate)
