--- AutoWrapper/AutoWrapper/CodeGen/DefaultContractNamingStrategy.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using AutoWrapper.CodeGen.Contracts;
-
-namespace AutoWrapper.CodeGen
-{
-	public class DefaultContractNamingStrategy : IContractNamingStrategy
-	{
-		public virtual string ContractNameFor(Type type)
-		{
-			return 'I' + type.Name;
-		}
-	}
-}
