+++ AutoWrapper/AutoWrapper/CodeGen/SuffixNamingStrategy.cs
@@ -3,11 +3,16 @@ using AutoWrapper.CodeGen.Contracts;
 
 namespace AutoWrapper.CodeGen
 {
-	public sealed class SuffixContractNamingStrategy : IContractNamingStrategy
+	public sealed class SuffixNamingStrategy : IContractNamingStrategy, ITypeNamingStrategy
 	{
 		public string ContractNameFor(Type type)
 		{
 			return $"I{type.Name}Wrapper";
+		}
+
+		public string TypeNameFor(Type type)
+		{
+			return $"{type.Name}Wrapper";
 		}
 	}
 }
